+++ aequilibrae/project/tools/network_simplifier.py
@@ -10,7 +10,7 @@ from shapely.ops import linemerge
 from shapely.ops import substring
 
 from aequilibrae.context import get_active_project
-from aequilibrae.traffic_assignment import Graph
+from aequilibrae.traffic_assignment.graph import Graph
 from aequilibrae.utils.aeq_signal import SIGNAL
 from aequilibrae.utils.db_utils import commit_and_close
 from aequilibrae.utils.interface.worker_thread import WorkerThread
