+++ aequilibrae/transit_assignment/optimal_strategies.py
@@ -6,7 +6,7 @@ from aequilibrae.paths.public_transport import HyperpathGenerating
 
 class OptimalStrategies:
     def __init__(self, assig_spec):
-        from aequilibrae.paths import TransitAssignment
+        from aequilibrae.transit_assignment.transit_assignment import TransitAssignment
 
         self.__assig_spec = assig_spec  # type: TransitAssignment
         self.__logger = assig_spec.logger
